refactored the organization baseball endpoints
