+++ app/core/urls/organization_urls.py
@@ -1,6 +1,7 @@
 from django.urls import path
 
 from ..views.organization_viewset import OrganizationViewSet
+from ..views.organization_baseball_viewset import OrganizationBaseballCompetitionViewset
 
 organization_list = OrganizationViewSet.as_view({
     'get': 'list',
@@ -13,15 +14,15 @@ organization_details = OrganizationViewSet.as_view({
     'delete': 'destroy'
 })
 
-organization_competition_list = OrganizationViewSet.as_view({
-    'get': 'list_competitions',
-    'post': 'create_competition'
+organization_competition_list = OrganizationBaseballCompetitionViewset.as_view({
+    'get': 'list',
+    'post': 'create'
 })
 
-organization_competition_details = OrganizationViewSet.as_view({
-    'get': 'retrieve_competition',
-    'patch': 'partial_update_competition',
-    'delete': 'destroy_competition'
+organization_competition_details = OrganizationBaseballCompetitionViewset.as_view({
+    'get': 'retrieve',
+    'patch': 'partial_update',
+    'delete': 'destroy'
 })
 
 organization_team_list = OrganizationViewSet.as_view({
@@ -38,8 +39,8 @@ organization_team_details = OrganizationViewSet.as_view({
 urlpatterns = [
     path('', organization_list),
     path('<uuid:organization_id>/', organization_details),
-    path('<uuid:organization_id>/competitions/', organization_competition_list),
-    path('<uuid:organization_id>/competitions/<uuid:competition_id>/', organization_competition_details),
-    path('<uuid:organization_id>/teams/', organization_team_list),
-    path('<uuid:organization_id>/teams/<uuid:team_id>/', organization_team_details),
+    path('<uuid:organization_id>/baseball/competitions/', organization_competition_list),
+    path('<uuid:organization_id>/baseball/competitions/<uuid:competition_id>/', organization_competition_details),
+    path('<uuid:organization_id>/baseball/teams/', organization_team_list),
+    path('<uuid:organization_id>/baseball/teams/<uuid:team_id>/', organization_team_details),
 ]
